.ref-proj/requirements.txt
typer
pydantic-ai
rich
python-dotenv
